+++ Data_science/DataMining/UMAP/Components/ProgressReporter.vb
@@ -42,4 +42,4 @@
 ''' <summary>
 ''' The progress will be a value from 0 to 1 that indicates approximately how much of the processing has been completed
 ''' </summary>
-Public Delegate Sub ProgressReporter(progress As Single)
+Public Delegate Sub ProgressReporter(progress As Double)
