+++ Data_science/DataMining/UMAP/Components/Tree/RandomProjectionTreeNode.vb
@@ -48,7 +48,7 @@ Namespace Tree
         Public Property Indices As Integer()
         Public Property LeftChild As RandomProjectionTreeNode
         Public Property RightChild As RandomProjectionTreeNode
-        Public Property Hyperplane As Single()
-        Public Property Offset As Single
+        Public Property Hyperplane As Double()
+        Public Property Offset As Double
     End Class
 End Namespace
